% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combination_with_repetition.R
\name{combination_with_repetition}
\alias{combination_with_repetition}
\title{Combination with repetition.}
\usage{
combination_with_repetition(n, r)
}
\arguments{
\item{n}{The number of elements (variables).}

\item{r}{The size of the groups (degreess of the polynomial interaction).}
}
\description{
From combinatorial math, this function aims calculates combinations with repetitions.
}
