% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estprod_data.R
\docType{data}
\name{estprod_data}
\alias{estprod_data}
\title{10000 randomly generated variables in panel data format.}
\format{A data frame with 10000 rows and 10 variables:
\describe{
  \item{id}{Identifies the 1000 randomly generated individuals.}
  \item{year}{The year associated to each individual observation.}
  \item{g1}{Put individuals in 25 groups.}
  \item{g2}{Put individuals in 50 groups.}
  \item{var1}{Randomly generated variable.}
  \item{var2}{Randomly generated variable.}
  \item{var3}{Randomly generated variable.}
  \item{var4}{Randomly generated variable.}
  \item{var5}{Randomly generated variable.}
  \item{exit}{The last year an id appears.}
}}
\usage{
estprod_data
}
\description{
10000 randomly generated variables in panel data format.
}
\keyword{datasets}
