% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{interactions}
\alias{interactions}
\title{Total number of \code{poly} elements.}
\usage{
interactions(n, d)
}
\arguments{
\item{n}{The number of variables.}

\item{d}{Degreess of the polynomial interaction.}
}
\description{
This function aims calculates the total number of terms of a polynomial interactions.
}
