% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_lag.R
\name{panel_lag}
\alias{panel_lag}
\title{Panel data lag function}
\usage{
panel_lag(x, id, time, lag = 1, verify = TRUE)
}
\arguments{
\item{x}{A vector, data.frame, tibble or matrix.}

\item{id}{A character with the name of the indicator variable.}

\item{time}{A character with the name of the time variable.}

\item{lag}{Number of lags.}

\item{verify}{Check if panel is sorted by id and time variables.}
}
\description{
This function aims create the lags of a specified variable from panel data.
}
\note{
Based on \href{https://bitbucket.org/paulschrimpf/}{Paul Schrimpf's} lag function.
}
