% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_elements.R
\name{poly_elements}
\alias{poly_elements}
\title{Number of \code{poly} elements.}
\usage{
poly_elements(n, d)
}
\arguments{
\item{n}{The number of variables.}

\item{d}{Degreess of polynomial interaction.}
}
\description{
This function aims calculate the number of terms of a polynomial interactions.
}
